Move ud(UserData) to ds(DocumentStore)
